+++ src/lib/server/socket_server.rs
@@ -21,9 +21,6 @@ use std::time::{Duration, Instant};
 
 use slog::Logger;
 
-/// The default port for server communication
-pub const DEFAULT_PORT: &str = "10842";
-
 /// Type for structs that interface with sockets
 pub trait SocketHandler {
     /// The type of the socket to use for sending/receiveing data