+++ src/lib/server/tcp.rs
@@ -7,7 +7,7 @@ use error::*;
 use message_handler::MessageHandler;
 use node::Node;
 use server::{Client, Server};
-use socket_server::{SocketClient, SocketHandler, SocketServer};
+use server::socket_server::{SocketClient, SocketHandler, SocketServer};
 
 use std::net::{IpAddr, Ipv6Addr, SocketAddr, TcpListener, TcpStream};
 use std::sync::Arc;