+++ src/lib/server/unix_socket.rs
@@ -10,7 +10,7 @@ use data_transformer::DataTransformer;
 use error::*;
 use message_handler::MessageHandler;
 use server::{LocalClient, LocalServer};
-use socket_server::{SocketHandler, SocketServer};
+use server::socket_server::{SocketHandler, SocketServer};
 use versioning;
 
 use std::fs;